mapbox_poi/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-mapbox-location-field==1.7.0
pytz==2020.5
sqlparse==0.4.1
